+++ src/output/file-write/absolute-conversion-export.js
@@ -89,5 +89,5 @@ function coordinateFileWrite(inpDestPth, inpGraphObj, fileHeadTxt, fwCallback)
 
 module.exports =
 {
-	performFileExport: performAbsoluteConversionFileExport
+	performExport: performAbsoluteConversionFileExport
 };