+++ src/output/file-write/graph-file-export.js
@@ -267,5 +267,5 @@ function compileBlankPath(gfsObject, compGraphObj, gText, fileResObj)
 
 module.exports =
 {
-	performGraphExport: performNodeGraphExport
+	performExport: performNodeGraphExport
 };