+++ src/output/file-write/raw-data-steps/path-data.js
@@ -318,9 +318,6 @@ function handleClose(dStream)
 }
 
 
-
-
-
 module.exports =
 {
 	createPathFile: createRawPathDataFile